Replacing all 'unsigned long' instances with 'size_t' and 'long' with 'sys::SSize_T'.  This fixed at least one bug that occurred on Windows in NITFReadControl::interleaved() with AOI's > 4 GB.  Also fixed a bug in test_dump_images when it was run with doing a single read - it was allocating and writing out more than was needed.  #306.
git-svn-id: https://wush.net/svn/mdaus-oss/trunk@485 9d772353-a6c9-4ee9-b2cc-df47a7649cd5

+++ modules/c++/six.sicd/include/six/sicd/RadarCollection.h
@@ -161,10 +161,10 @@ struct AreaDirectionParameters
     double spacing;
 
     //!  The number of elements in this direction (SICD NumLines/NumSamples)
-    unsigned long elements;
+    size_t elements;
 
     //!  The first element in this direction (SICD FirstLine/FirstSample)
-    unsigned long first;
+    size_t first;
 };
 
 /*!
